requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
segno==1.5.2
